1.basic file restructing(renaming, deleting, rebasing files) 2. Basic css to make footer as well as card display work(_layout plus extra older version of bootstrap)(css priority ensures that we only load the older features when they are ommited from the newer version and not vice versa) 3. Made deserializer a service so it can be injected to anyone that needs it( now he loads posts at every call and does it pretty fast)(in the future that should probably be changed)(perhaps find a way to deserialize xml + paging?)(note : EF framework when we move to DB supports that easily) 4. Add basic post list + post layout 5. Changed the default home page(goes to postList by default)
