+++ Pages/Post.cshtml
@@ -1,0 +1,15 @@
+@page
+@model PostModel
+@{
+    ViewData["Title"] = "Post";
+}
+
+<div class="card">
+    <div class="card-body">
+        <h1 class="card-title">@Model.post.Title</h1>
+        <em><h3>by @Model.post.Author</h3></em>
+        <h4>on @Model.post.Date.ToString("dd-MM-yy HH:mm:ss")</h4>
+        <a href="@Model.post.Link">Original thread</a>
+        <pre class="card-text">@Html.Raw(Model.post.Body)</pre>
+    </div>
+</div>
