+++ Posts/posts.xsd
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<xs:schema xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xs="http://www.w3.org/2001/XMLSchema" attributeFormDefault="unqualified" elementFormDefault="qualified">
+    <xsd:element name="posts">
+        <xsd:complexType>
+            <xsd:sequence>
+                <xsd:element maxOccurs="unbounded" name="post">
+                    <xsd:complexType>
+                        <xsd:sequence>
+                            <xsd:element name="author" type="xsd:string" />
+                            <xsd:element name="date" type="xsd:dateTime" />
+                            <xsd:element name="link" type="xsd:string" />
+                            <xsd:element name="title" type="xsd:string" />
+                            <xsd:element name="body" type="xsd:string" />
+                        </xsd:sequence>
+                    </xsd:complexType>
+                </xsd:element>
+            </xsd:sequence>
+        </xsd:complexType>
+    </xsd:element>
+</xs:schema>
