+++ Utilities/Interfaces/IPostLoader.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using QCUtilities.Entities;
+
+namespace QCUtilities.Interfaces
+{
+    public interface IPostLoader
+    {
+        public List<Post> DeserializeXML(string fileName,string xsd);
+
+    }
+}
